maker/pyproject.toml
[tool.poetry]
name = "consistency-dockerfile-maker"
version = "0.0.1"
description = "A consistency script to render Dockerfile from templates."
authors = ["François Michaud <fmichaud@deepnox.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
PyYAML = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

